+++ compiler-core/src/type_/tests.rs
@@ -886,6 +886,7 @@ fn infer_module_type_retention_test() {
             type_aliases: HashMap::new(),
             documentation: Vec::new(),
             contains_echo: false,
+            references: References::default()
         }
     );
 }
@@ -2741,6 +2742,7 @@ fn assert_suitable_main_function_not_module_function() {
                 can_run_on_erlang: true,
                 can_run_on_javascript: true,
             },
+            name: "main".into(),
         },
     };
     assert!(assert_suitable_main_function(&value, &"module".into(), Target::Erlang).is_err(),);