+++ CHANGELOG.md
@@ -298,6 +298,40 @@
   name.
   ([Surya Rose](https://github.com/GearsDatapacks))
 
+- The Language Server now allows renaming of functions, constants and type variants
+  across modules. For example:
+
+  ```gleam
+  // wibble.gleam
+  pub fn wibble() {
+    wibble()
+  //^ Trigger rename
+  }
+  // wobble.gleam
+  import wibble
+
+  pub fn main() {
+    wibble.wibble()
+  }
+  ```
+
+  Becomes:
+
+  ```gleam
+  // wibble.gleam
+  pub fn wobble() {
+    wobble()
+  }
+  // wobble.gleam
+  import wibble
+
+  pub fn main() {
+    wibble.wobble()
+  }
+  ```
+
+  ([Surya Rose](https://github.com/GearsDatapacks))
+
 ### Formatter
 
 - Redundant function captures that take no additional arguments are now